+++ src/features/progress/models/index.ts
@@ -1,1 +1,2 @@
 export type { ProgressConfig, ProgressInspectionConfig } from './progress.model';
+export type { DrinoProgressEvent } from './event-progress.model';
