+++ src/request/models/index.ts
@@ -1,2 +1,1 @@
 export type { RequestConfig } from './request-config.model';
-export type { DrinoProgressEvent } from './request-controller.model';
